feat(docs): reorganize documentation structure and fix build errors
Documentation reorganization:
- Move implementation plans to docs/plans/
  * IMPLEMENTATION_PLAN_RPI5_HARDWARE.md
  * IMPLEMENTATION_PLAN_SISLANG.md
  * IMPLEMENTATION_PLAN_STUBS_TO_REAL.md
  * IMPLEMENTATION_PLAN_X86_64.md
- Create docs/implementation/rpi5/ directory
- Move all RPI5 documentation to docs/implementation/rpi5/
  * RPI5_HARDWARE_IMPLEMENTATION.md
  * RPI5_IMPLEMENTATION_SUMMARY.md
  * RPI5_INTEGRATION_GUIDE.md
  * RPI5_M1_STORAGE.md
  * RPI5_M2_POWER.md
  * RPI5_M3_SMP.md
  * RPI5_M7_VALIDATION.md
  * RPI5_M8_HARDENING.md
  * RPI5_M8_HARDENING_SUMMARY.md
  * RPI5_REVIEW_M0_M1.md

Build fixes:
- Add parse_number() helper to shell.rs for hex/dec parsing
- Fix compilation errors across RPI5 driver modules
- Update imports and error handling in platform/validation code

README updates:
- Update phase status descriptions from 'PRODUCTION READY' to
  'Implemented and exercised in QEMU' for accuracy

+++ crates/kernel/src/shell/gpio_helpers.rs
@@ -41,7 +41,7 @@ impl super::Shell {
                     match crate::drivers::gpio::set_pin(pin as u32) {
                         Ok(()) => unsafe {
                             crate::uart_print(b"[GPIO] Pin ");
-                            self.print_number_simple(pin);
+                            self.print_number_simple(pin as u64);
                             crate::uart_print(b" set HIGH\n");
                         },
                         Err(e) => self.print_gpio_error(e),
@@ -57,7 +57,7 @@ impl super::Shell {
                     match crate::drivers::gpio::clear_pin(pin as u32) {
                         Ok(()) => unsafe {
                             crate::uart_print(b"[GPIO] Pin ");
-                            self.print_number_simple(pin);
+                            self.print_number_simple(pin as u64);
                             crate::uart_print(b" set LOW\n");
                         },
                         Err(e) => self.print_gpio_error(e),
@@ -73,7 +73,7 @@ impl super::Shell {
                     match crate::drivers::gpio::toggle_pin(pin as u32) {
                         Ok(()) => unsafe {
                             crate::uart_print(b"[GPIO] Pin ");
-                            self.print_number_simple(pin);
+                            self.print_number_simple(pin as u64);
                             crate::uart_print(b" toggled\n");
                         },
                         Err(e) => self.print_gpio_error(e),
@@ -89,7 +89,7 @@ impl super::Shell {
                     match crate::drivers::gpio::read_pin(pin as u32) {
                         Ok(level) => unsafe {
                             crate::uart_print(b"[GPIO] Pin ");
-                            self.print_number_simple(pin);
+                            self.print_number_simple(pin as u64);
                             crate::uart_print(b" = ");
                             if level {
                                 crate::uart_print(b"HIGH\n");
@@ -110,7 +110,7 @@ impl super::Shell {
                     match crate::drivers::gpio::set_function(pin as u32, GpioFunction::Output) {
                         Ok(()) => unsafe {
                             crate::uart_print(b"[GPIO] Pin ");
-                            self.print_number_simple(pin);
+                            self.print_number_simple(pin as u64);
                             crate::uart_print(b" configured as OUTPUT\n");
                         },
                         Err(e) => self.print_gpio_error(e),
@@ -126,7 +126,7 @@ impl super::Shell {
                     match crate::drivers::gpio::set_function(pin as u32, GpioFunction::Input) {
                         Ok(()) => unsafe {
                             crate::uart_print(b"[GPIO] Pin ");
-                            self.print_number_simple(pin);
+                            self.print_number_simple(pin as u64);
                             crate::uart_print(b" configured as INPUT\n");
                         },
                         Err(e) => self.print_gpio_error(e),
@@ -148,9 +148,9 @@ impl super::Shell {
 
                     unsafe {
                         crate::uart_print(b"[GPIO] Blinking pin ");
-                        self.print_number_simple(pin);
+                        self.print_number_simple(pin as u64);
                         crate::uart_print(b" ");
-                        self.print_number_simple(count);
+                        self.print_number_simple(count as u64);
                         crate::uart_print(b" times...\n");
                     }
